fix: merge conflicts and task row pr integration
