+++ frontend/src/community/crm/api/utils/ApiEndpoints.ts
@@ -1,9 +1,5 @@
 import { moduleAPIPath } from "~community/common/constants/configs";
 
-export const taskEndpoints = {
-  UPDATE_TASK: (id: number) => `${moduleAPIPath.CRM}/task/${id}`
-};
-
 export const crmDealEndpoints = {
   GET_DEALS: `${moduleAPIPath.CRM}/deal`,
   CREATE_DEAL: `${moduleAPIPath.CRM}/deal`,
@@ -20,6 +16,7 @@ export const contactEndpoints = {
 };
 
 export const taskEndpoints = {
+  UPDATE_TASK: (id: number) => `${moduleAPIPath.CRM}/task/${id}`,
   GET_OPEN_TASKS: `${moduleAPIPath.CRM}/task`
 };
