+++ frontend/src/community/crm/components/atoms/TaskGroup/TaskGroup.tsx
@@ -1,8 +1,7 @@
 import { FC } from "react";
 
 import { CrmTaskDetailType } from "~community/crm/types/CommonTypes";
-
-// import TaskRow from "../TaskRow/TaskRow";
+import TaskRow from "../../molecules/TaskRow/TaskRow";
 
 interface TaskGroupProps {
   label: string;
@@ -17,7 +16,7 @@ const TaskGroup: FC<TaskGroupProps> = ({ label, tasks }) => {
       <div className="subtitle2 mb-2 sticky top-0 bg-white">{label}</div>
       <div className="border border-secondary-accent rounded-lg overflow-hidden divide-y divide-secondary-accent">
         {tasks.map((task) => (
-          <div key={task.id}>{/* <TaskRow task={task} /> */}</div>
+          <div key={task.id}><TaskRow task={task} /></div>
         ))}
       </div>
     </div>
