+++ frontend/src/community/crm/components/molecules/OpenTasksTabContent/OpenTasksTabContent.tsx
@@ -14,7 +14,7 @@ import { useGetOpenTasks } from "~community/crm/api/TaskApi";
 import { TASK_SEARCH_DEBOUNCE_DELAY } from "~community/crm/constants/taskConstants";
 import { CrmTaskTabEnum } from "~community/crm/enums/common";
 import { getEmptyStateType } from "~community/crm/utils/commonHelpers";
-import { getTaskGroups } from "~community/crm/utils/taskUtils";
+import { getTaskGroups } from "~community/crm/utils/taskUtil";
 
 import TaskGroup from "../../atoms/TaskGroup/TaskGroup";
