+++ frontend/src/community/crm/components/molecules/TaskRow/TaskRowContent.tsx
@@ -1,13 +1,13 @@
 import { FC } from "react";
 
-import { CrmTaskType } from "~community/crm/types/CommonTypes";
+import { CrmTaskDetailType } from "~community/crm/types/CommonTypes";
 import { getTaskTypeIcon } from "~community/crm/utils/taskUtil";
 
 import TaskRowMeta from "./TaskRowMeta";
 import TaskRowSubtitle from "./TaskRowSubtitle";
 
 interface Props {
-  task: CrmTaskType;
+  task: CrmTaskDetailType;
   isShowContact: boolean;
   applyCompletedStyle: boolean;
 }
@@ -22,7 +22,7 @@ const TaskRowContent: FC<Props> = ({
       <div
         className={`shrink-0 flex items-center justify-center ${applyCompletedStyle ? "opacity-40" : ""}`}
       >
-        {getTaskTypeIcon(task.type.name)}
+        {getTaskTypeIcon(task.typeName)}
       </div>
 
       <div className="flex-1 min-w-0">
