+++ frontend/src/community/crm/components/molecules/TaskRow/TaskRowSubtitle.tsx
@@ -1,11 +1,11 @@
 import { FC } from "react";
 
 import { useTranslator } from "~community/common/hooks/useTranslator";
-import { CrmTaskType } from "~community/crm/types/CommonTypes";
+import { CrmTaskDetailType } from "~community/crm/types/CommonTypes";
 import { getDueDateStatus } from "~community/crm/utils/taskUtil";
 
 interface Props {
-  task: CrmTaskType;
+  task: CrmTaskDetailType;
   isShowContact: boolean;
   applyCompletedStyle: boolean;
 }
